Fixed a few TODOs.
Summary:
1) Moved jshint to a third-party directory

2) Moved calculator.js to genfiles.  This is because I've noticed
several commits over the years that have edited calculator.js
directly, only to have that edit overwritten the next time it was
auto-regenerated.  By putting it in genfiles, I hope to indicate to
people that they should change the source .jison file instead.  (I
still check in the genfiles/ file, so folks can use khan-exercises
without needing to install jison.)

While testing, I realized that we don't do a very good job of
installing deps for you, so I added deps-manifests for the three (!)
programming languages we use, and changed the Makefile to use them.

Small cleanups based on review.

Test Plan:
To test the jshint move, I ran:
   env LC_ALL=en_US.UTF-8 .bundle exec ruby build/pack.rb exercises/z_scores_3.html::exercises-packed/z_scores_3.html
without error.

To test the calculator move/change, I ran:
   python -mSimpleHTTPServer
and visited
   http://localhost:8000/exercises/law_of_cosines.html
and clicked '3+4=' and got 7, and clicked 'ln(-5)' and saw the bar
turn red because it's an error.  (It looks like the actual error
message is suppressed these days.)

(same)

Reviewers: eater, alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D9356