Get rid of pluralTex.
Summary:
Ran
   grep pluralTex -r *
and replaced the one place where it was found in the code with the
appropriate expansion.

Test Plan:
Ran
   build/lint_i18n_strings_test.py

Then ran
   python -mSimpleHTTPServer
and visited
   http://localhost:8000/exercises/place_value.html
until it said
   "What is the place value of 4 in 6314?"
and then I clicked on 'like a hint' until it said:
   = 6 thousands + 3 hundreds + 1 ten + 4 ones

which shows the correct pluralization, and the 6/3/1/4 were in latex
font like they should be (while the thousands/hundreds/etc weren't).

Reviewers: eater, john

Reviewed By: john

Differential Revision: http://phabricator.khanacademy.org/D9358