Make the ruler widget more intuitive to use.
Summary:
Currently, dragging the body of the ruler rotates it around the
point on the bottom left, and dragging the point translates the ruler.
This change makes dragging the body of the ruler translate it, and puts
points at the bottom left and right of the ruler. Dragging these points
rotates the ruler about the opposite point.

Test Plan:
- Opened test.html and added a measurer widget
- Played with ruler (translation, rotation with either point)

Reviewers: jack, joel

Reviewed By: joel

Subscribers: charlie

Differential Revision: http://phabricator.khanacademy.org/D9625