Fail loudly when makeProblem is called with bad exerciseId
Summary:
I ran into this because the server was returning a blank response instead of a
real 404. I fixed that in rGWb80e1799732445d4a1de45fcd3e14b4c6e2ad462, but to
make it easier to debug things like this in the future, it's better if
`makeProblem` fails loudly instead of silently failing to render anything.

Test Plan: Load my broken pretest task. See an error in the console.

Reviewers: alpert, eater

Reviewed By: eater

Differential Revision: http://phabricator.khanacademy.org/D9666