Better event handling for movablePoint and mouseLayer.
Summary:
In order to have "Click to add points" create the new point on mouse down
and demonstrate the correct drag behavior, some of the event handling in interactive
needed to be refactored. The mouseLayer now supports a wider range of mouse events
(and binds mouse move events to $(document) to provide identical behavior on touch
and browser devices) and movablePoint now has a "grab" method to allow for initial
grabbing.

Test Plan: Same as for D9434

Reviewers: eater, jack

Reviewed By: jack

Differential Revision: http://phabricator.khanacademy.org/D9436