Assign Khan, KhanUtil earlier
Summary: webpack executes async requires synchronously when the required module is in the same chunk, which throws us off here if DOM-ready has already happened when this package loads because answer-types (and probably others) get loaded synchronously by the `loadModule` call and expect to see `Khan` in the global scope. This is maybe a bug in webpack, but we can work around it pretty easily.

Test Plan: Load metadash without errors.

Reviewers: jack

Reviewed By: jack

Differential Revision: http://phabricator.khanacademy.org/D11357