Fix calculator so it always shows the large output window (Fixes T2570)
Summary:
We want to always show the full calculator output window on calculator
questions, and we also want to clear the history from one question to
the next. Removing `history`'s children unfortunately removed the output
window as well - what we really want to do is just remove the history
*lines* from inside the `#calc-output-content` div.

Test Plan:
Opened an exercise with a calculator (e.g. /exercises/z_scores_2.html)
Observed that the calculator behaved normally on the first question
Proceeded to the next question, and observed that:
  1. the large output window now appears
  2. the calculator's history was cleared to a blank state
Proceeded to subsequent questions and continued to observe good things

Reviewers: jack

Reviewed By: jack

Subscribers: david, charlie, eater, jack

Maniphest Tasks: T2570

Differential Revision: http://phabricator.khanacademy.org/D11327