division_1 hints no longer modify graphie in the question-area (for first problem-type)
Summary:
This was leading to some pain on mobile, where the hints are strictly below
the answer-area (and thus detached from the question area). Now, the hints use their
own separate graphie. (I also modified one of the hints, which didn't seem to be
associated with the correct problem type.)

Test Plan:
- Run khan-exercises.
- Run "division_1".
- Check hints for each problem type with a variety of values.
- Verify that the hints don't run into the answer-area (i.e., are too long).
- Verify that the hints don't get cut off on the left.
- Verify that nothing looks out-of-place!

Reviewers: eater

Reviewed By: eater

Subscribers: marcos

Differential Revision: http://phabricator.khanacademy.org/D11808