Automatically cancel graphie intervals during cleanup
Test Plan:
Go to "Volume with Fractions".
Profile with developer tools.
Note that a timer is firing every 50ms.
Solve the problem without touching the figure.
Go to the next problem.
Touch the figure to cancel its timer.
Profile with developer tools.
Assert that no timers are firing
(previously, the timer from the first problem
would continue to fire).

Reviewers: alpert

Reviewed By: alpert

Subscribers: emily

Differential Revision: http://phabricator.khanacademy.org/D12658