Add labels to inscribed circles diagram
Summary:
Originally this was to add labels to the angles so they're not just refered to by colour. But I also noticed the hints referred to previous exercise, which doesn't make sense now they're all one exercise. The hints for the last two exercise types are a bit involved, but hopefully after a couple of exercises people will get the idea that the blue angle is twice the orange one.

Update hints to use coloured angles

Update hints

Make hints more consistent

Avoid reference to other exercises in hints

Finalise hints

Reviewers: eater

Reviewed By: eater

Differential Revision: http://phabricator.khanacademy.org/D12809