Update validator so non-integers can be used
Summary: This is the exercise which I based the point-slope form validator on. I've updated it to be more like the validator you wrote. I can't help but feel there must be an easier, less repetitive way. The exercise now works for non-integers which is something I have wanted to change for a while.

Reviewers: emily

Reviewed By: emily

Differential Revision: http://phabricator.khanacademy.org/D13062