Change khan-exercises to use /api/internal everywhere
Summary:
The new convention is that the calls made by the website use /api/internal, so
it's clear that we have more control over them.

The old and new names are already both valid, so the timing of this commit
isn't fragile.

Test Plan: Do some exercises, take some hints.

Reviewers: eater

Reviewed By: eater

Subscribers: csilvers

Differential Revision: http://phabricator.khanacademy.org/D13703