Add graphie.parabola function to plot parabolas using SVG quadratic paths
Summary:
SVG Biezer paths allow for a custom quadratic curve command that's much more
performant than creating an massive path string using thousands of points, as we were
doing before. I added a graphie.parabola function that takes the parabola's vertex
and any other point on the parabola as arguments. Note that the range of the plotted parabola
is based on the x-distance between the vertex and the provided point.

To demonstrate how this function will be used, I also converted graphing_parabolas_1.html to use
this new function call. I don't have my iPad with me, but I recall the proof-of-concept version
of this same implementation yielding a 2x+ FPS improvement.

Once we've accepted this review, I can convert the other khan-exercises as well with relative ease.

Test Plan:
- Run khan-exercises.
- Open graphing_parabolas_1.html.
- Move the points around and verify that behavior is unchanged.
- Verify that grading works.

Reviewers: aria

Reviewed By: aria

Subscribers: eater

Differential Revision: http://phabricator.khanacademy.org/D13977