Fix typo and expand colour coding
Summary:
For once some typos that aren't my fault. Change "in" to "is" and add another "in". I move some of the data-ensures so they are only on the radius variables as they're all that matter. If the radius is <= 10, then the real and imaginary parts will between -10 and 10 anyway (unless I'm being stupid).

Since we're using coloured dots, I thought we might as well colour the numbers too. I've removed the style elements since they exist in khan-exercises.css already.

Reviewers: eater, emily

Reviewed By: emily

Differential Revision: http://phabricator.khanacademy.org/D13789