Replace coordinates with labels and colour-code hints
Summary:
Replacing the coordinates with labels so they are more likely to be easier to read.
For Elizabeth: https://khanacademy.atlassian.net/browse/AI-382585
This also means the questions can be rewritten so hopefully fewer people will try to find the midpoint of the given points in the second exercise type.

Reviewers: eater, emily

Reviewed By: emily

Differential Revision: http://phabricator.khanacademy.org/D13864