Move all answer area content to the bottom of the question area
Summary:
This commit moves the entire #solutionarea div to the bottom of the
question area, unless the question contains an element with the class
name "render-answer-area-here" (in which case the div is placed
immediately after this element).

In addition, this commit removes entirely the old "acceptable formats"
section at the top of the answer area in favor of tooltips attached
directly to individual input elements. This is basically backporting the
approach we've taken in Perseus to khan-exercises. Speaking of which,
this commit also brings the expression editor in khan-exercises more in
line with the Perseus version (rendered output above input) which is a
necessary change because otherwise the tooltip obscures the output.

Test Plan:
Most of these changes were in place during the answer area moving party,
and so in general all exercises should be confirmed to work. Tooltips
were only finished afterwards though, so I've spot-checked the various
answer types myself.

Reviewers: emily

Subscribers: aria

Differential Revision: http://phabricator.khanacademy.org/D14745