Replace data-weight=0 template problem type with proper templating
Summary:
khan-exercises already supports inheriting from a template: the stuff that's
not part of a problem type. Doing it this way means there isn't a "template"
problem type sitting around that the screenshotting thing tries and fails to
take a screenshot of.

Test Plan:
Hit all of these and click through the different problem types to see that they result in a reasonable-looking question. Take a few hints here and there and verify they make sense:
http://exercises.ka.local/exercises/dice_probability.html?debug
http://exercises.ka.local/exercises/ideal_stoichiometry.html?debug
http://exercises.ka.local/exercises/limiting_reagent_stoichiometry.html?debug
Note the lack of a "template" problem type in any of these that results in JS errors.

Reviewers: emily

Reviewed By: emily

Differential Revision: https://phabricator.khanacademy.org/D14869