Make balancing_chemical_equations work better
Summary: It looks like, before we merged the answer area into the
problem, we wanted to have inline inputs inside the question for this
problem, so we made a custom answer type with a custom validator.
However, this causes a problem for mobile which doesn't discover the
input boxes we created (and make them pretty and use the MyScript
input). Instead of making mobile smarter and discovering the input
boxes, I just redid the exercise to do things using normal multiple
answer types, since the answer area is now put into the body of the
question. (was going to be reviewed, but pushed up for mobile reasons:
https://phabricator.khanacademy.org/D15505)

Test Plan:
 - Go through each problem type:
   - Ensure that the correct solution lets you get the problem correct.
   - Ensure that leaving boxes where the solution is '1' empty still
     grades correctly.

Auditors: eater