Wrap khan-exercises' MovablePoints in divs
Summary:
Similar to how we wrap Perseus' MovablePoints in divs, this brings the same functionality
and performance boost to khan-exercises.

My main point of contention with this diff is that there's a lot of code reuse (copying of
files, etc.). It's not entirely clear to me if there's a better way to share code between Perseus
and khan-exercises.

In addition, I ran into some not-niceties with animations. In khan-exercises, we sometimes
animate points to new coordinates, which didn't work perfectly out-of-the-box. There's now a new
`animateTo` method for WrappedPoints that animates appropriately. I grepped around the codebase
places in which we currently have `point.animate(// new coord)` calls (to replace them with calls
to `animateTo`), but couldn't find any beyond the one in interactive.js. In general, there's a lot
of weird behavior in khan-exercises that makes me nervous about pushing this, but testing various
exercises renewed my confidence.

Test Plan:
- Run khan-exercises.
- To test basic functionality, try [The Complex Plane](http://localhost:8000/exercises/the_complex_plane.html).
    - Verify that the point grows on mouse-over.
    - Verify that you can click and drag, but not off of the grid.
    - Verify that the point snaps to grid.
- To test animations, try [Decimals on the Number Line 2](http://localhost:8000/exercises/decimals_on_the_number_line_2.html).
    - Cycle through all the hints. The last hint should animate the point to the end of the new blue line.
- To test animations with synced lines, try [Graphing Systems of Equations](http://localhost:8000/exercises/graphing_systems_of_equations.html).
    - Cycle through all the hints. The last hint should animate the points and lines in sync.
- To test advanced behavior, try [Congruency Postulates](http://localhost:8000/exercises/congruency_postulates.html).
    - Verify that the basic functionality is preserved.
    - Verify that the "Reflect Shape" button reflects the shape.

Reviewers: aria, emily

Reviewed By: emily

Subscribers: andy

Differential Revision: https://phabricator.khanacademy.org/D15452