Add event after initial hints are rendered
Summary:
We have events bound to the Exercises object that are triggered
after the main body of the problem is rendered but not the hints, and
events that trigger after each additional hint is taken, but no event
that is triggered after the initial hints are rendered. This adds such
an event, so that mobile can modify the initial hints (in particular, so
it can replace qhint text boxes with myscript ones).

Test Plan:
 - Put this code in the mobile app
 - Bind to this event and put a breakpoint
 - Load an exercise where hints have been taken
 - See that the hints are part of the body of the question when the
   breakpoint is hit

Reviewers: joel

Reviewed By: joel

Differential Revision: https://phabricator.khanacademy.org/D15548