Remove places khan-exercises manually sets font-size
Summary: In a bunch of places, we manually set the font-size (in pixels)
of certain parts of the question (usually in the answer area so it would
fit in the small box). This is unnecessary now since the answer area is
in the body, and it looks bad on mobile because the fonts don't scale
accordingly. This fixes a couple places that this happened.

Also, I fixed the TeX in parabola_intuition_{2,3}, because it was a
mess.

Test plan:
 - Open all the exercises that changed locally, see that they look good
 - Open all the exercises that changed on mobile, see that they look
 good

Auditors: joel, eater