Check Answer: Remove defunct "Try Again" code
Summary:
This code currently doesn't do anything useful.
We can revisit this in the future if we want to (though
we all seem to sort of like the current way), but I guess
for now the code ought to reflect what actually happens ;)

Test Plan:
Do additon_1 in webapp and see that I can still check answer
and progress.

Auditors: emily, eater