Fix views_of_a_function answer type in iPad app
Summary:
In the iOS app, we hide the inputs and show MathOutput thingies instead. This
means that we muck around in the DOM such that each <td> has a <span> and a
hidden <input> as children rather than just the <input>. The guess extraction
part of the custom answer type assumed that the <input>s were the only children
of <td>s. This change explicitly looks for input elements rather than assuming.

This also adds a fix that prevents grading if you don't fill out any answers.

Test Plan:
iOS
===
1. Update khan-exercises in the iOS app with this change.
2. Run the iPad app, find the Views of a function exercise.
3. Answer a question that requires filling out a table, submit and get it right.
4. Hit "check" on a question that requires filling out a table without entering anything and see that it doesn't grade it.

Local mode
===
1. Load http://exercises.ka.local/exercises/views_of_a_function.html
2. Answer a question that requires filling out a table, submit and get it right.
3. Hit "check answer" on a question that requires filling out a table without entering anything and see that it doesn't grade it.

Reviewers: emily

Reviewed By: emily

Differential Revision: https://phabricator.khanacademy.org/D16426