Send /attempt and /hint requests to a multithreaded module when possible.
Summary:
Multithreaded modules (such as 'batch') are slower than the default
module, but cheaper.  We use it when we don't mind the extra time
taken.  In this case, we use it inside practice tasks, but not at the
end of a task, because we know it will take a few seconds to finish
the next problem in the task, so it doesn't matter if this API request
takes a few seconds as well.  (Inside a practice task, the return
value of the API call doesn't matter that much.)

Test Plan:
I ran 'make serve' and visited
      http://localhost:8080/math/geometry/intro_euclid/e/drawing-lines
and clicked 'hint' and then answered the question right, and clicked
'next question', and saw on the chrome network tab that it was hitting
/_mt/ in each case, and there were no javascript or other errors.

But the real test will be to do this in prod, when the batch module
will be quite a bit slower than the default module (unlike in dev).
Fingers crossed on that.

Reviewers: alan

Reviewed By: alan

Subscribers: desmond

Differential Revision: https://phabricator.khanacademy.org/D16572