Fix keyboard navigation and provide alerts when attempting to complete problems. Fixes T3679.
Summary:
This is only for the exercise framework, not for Perseus (that'll be another diff).

This makes it so that when you attempt to complete a problem and it is incorrect it'll read out a message explaining what went wrong (currently what is shown as the 'attempt message' beneath the button). It also stops the button from wiggling as there is now a textual message explaining that. Note the comments inline about these changes.

Test Plan: It should now be possible, using a screen reader, to successfully navigate an exercise and at least submit an answer (and be told when you get that answer wrong).

Reviewers: aria, jordan

Reviewed By: jordan

Maniphest Tasks: T3624

Differential Revision: https://phabricator.khanacademy.org/D16868