Log on /attempt-sending error in all cases, not just timeouts.
Summary:
This will, I'm hoping, help me debug the problem-out-of-order error.
I have a theory that one cause is when requests don't make it to the
server because the queue is cleared.  This will provide data to
possibly prove or disprove the theory.

The url being logged previously was not the full url, so I
restructured the code some so we do get the full url.  I also use a
different hack than before to try to distinguish navigating-away from
unresponsive server.

Test Plan:
Ran 'make serve' and visit
   http://localhost:8080/math/geometry/intro_euclid/e/drawing-lines
and answered the question right.  I saw /attempt return a 200 (in the
appengine logs) and the smiley face show up on the webpage.  I then
answered the next question wrong and again saw a 200 and a reasonable
UI.

To test handling of a server error, I then modified
interface.js:fullUrl to add this line right before the return:
    apiBaseUrl = 'http://localhost:8888' + apiBaseUrl;

Then, in a shell, I ran this code to make a balky server:
   python -c 'import socket; a = socket.socket(2,1); a.bind(("", 8888)); a.listen(5); b = a.accept()'

I then ran 'make serve' again, and revisited
   http://localhost:8080/math/geometry/intro_euclid/e/drawing-lines
I answered the question right, and immediately saw the 'out of date'
message.  I also saw, in the logs:
  INFO     2015-04-08 19:46:02,602 main.py:175] Manually sent to log: [1428522362556] request to http://localhost:8888/api/internal/_mt/user/exercises/drawing-lines/problems/4/attempt failed (0, error) with 0 other pending API requests (in khan-exercises/interface.js:handleAttempt) (user email = None)

I then ran a different python server:
   python -c 'import socket, time; a = socket.socket(2,1); a.bind(("", 8888)); a.listen(5); b = a.accept(); time.sleep(1000)'

I then revisited
   http://localhost:8080/math/geometry/intro_euclid/e/drawing-lines
I answered the question right, then closed the tab.  Nothing was
logged, even afer 30 seconds.

Finally, to test the timeout code, I ran the same python server again,
revisited
   http://localhost:8080/math/geometry/intro_euclid/e/drawing-lines
and answered the question right, and waited.  After 30 seconds, I saw
in the logs:
   INFO     2015-04-08 19:52:52,508 main.py:175] Manually sent to log: [1428522772454] request to http://localhost:8888/api/internal/_mt/user/exercises/drawing-lines/problems/5/attempt failed (0, timeout) with 0 other pending API requests (in khan-exercises/interface.js:handleAttempt) (user email = None)

Reviewers: david, emily

Reviewed By: emily

Subscribers: emily, tom

Maniphest Tasks: T3704

Differential Revision: https://phabricator.khanacademy.org/D17274