Log more information on out-of-order (and other) errors.
Summary:
Before we logged the length of the request queue when we saw an
error.  Now we log the actual contents (urls) of the queue as well.
That may help me track down exactly what's going on when we see an
out-of-order error.

Test Plan:
Ran
   make serve
and visited
   http://localhost:8080/math/algebra-basics/core-algebra-expressions/core-algebra-variables-and-expressions/e/evaluating_expressions_2
and asked for a hint, then answered a question right.  No errors in
that.

To test handling of a server error, I then modified
interface.js:fullUrl to add this line right before the return:
    apiBaseUrl = 'http://localhost:8888' + apiBaseUrl;

Then, in a shell, I ran this code to make a balky server:
   python -c 'import socket; a = socket.socket(2,1); a.bind(("", 8888)); a.listen(5); b = a.accept()'

I then ran 'make serve' again, and revisited
   http://localhost:8080/math/algebra-basics/core-algebra-expressions/core-algebra-variables-and-expressions/e/evaluating_expressions_2
I answered the question right, and immediately saw the 'out of date'
message.  I also saw, in the logs:
   INFO     2015-04-29 19:10:15,498 main.py:175] Manually sent to log: [1430334615465] request to http://localhost:8888/api/internal/_mt/user/exercises/evaluating_expressions_2/problems/3/attempt failed (0, error) with 0 other pending API requests: http://localhost:8888/api/internal/_mt/user/exercises/evaluating_expressions_2/problems/3/attempt (in khan-exercises/interface.js:handleAttempt) (user email = None)

so it looks like the logging-of-urls is happening correctly, at least
in this simple case.

Reviewers: david, emily

Reviewed By: emily

Subscribers: alpert

Differential Revision: https://phabricator.khanacademy.org/D17603