Rewrite pack.rb in javascript, to remove the main ruby dep on our codebase.
Summary:
With this, it's practical to remove the need for ruby in our codebase
(a few other tools use it, but those tools don't seem to be in active
use).

The code is more natural in node.js anyway, since a lot of the
libraries used are ones that were written for javascript and then
ported to ruby.

The html normalizer here works slightly differently than the ruby one,
making the diff annoying to do.  But what I did was ran

   % rm exercises-packed/* && make pack
   % rsync -av exercises-packed/ exercises-packed.old/

before making my changes, and then after the changes ran

   % rm exercises-packed/* && make pack
   % diff -ru exercises-packed.old exercises-packed | grep -v '&quot' | grep -v '</p><p data-' | grep -v '=""' | less

The greps are two avoid three major classes of changes:
 1) the new prettifier changes " to &quot; all over the place
    (annoying, but oh well)
2) the new prettifier doesn't put in newlines in as many places as the
   old one did
3) the new prettifier replaces 'raw' attributes with 'attribute=""'
   (should be harmless).

I then looked for all the '+' lines in the resulting diff, and
manually examined them.  The major remaining differences were:

1) The old prettifier left in utf-8 literals.  The new one uses html
   entities.
2) The new code uses a more recent version of the uglifier, which
   uglifies code slightly differently (but, I presume,
    equivalently)

Test Plan:
Ran
   python -m SimpleHTTPServer
and visited these pages:
   http://localhost:8000/exercises-packed/absolute_value.html
   http://localhost:8000/exercises-packed/dividing_decimals_0.5.html
   http://localhost:8000/exercises-packed/hyperbola_intuition.html

In each case I looked at all the hints then answered the question
correctly, with success.

Reviewers: eater

Reviewed By: eater

Differential Revision: https://phabricator.khanacademy.org/D17606