Fix the regex intended to strip external scripts
Summary:
We have to strip out external scripts when loading the exercise html file,
otherwise, IE will try to execute them and try to fetch local-only files for
which we return 404s.

See https://github.com/Khan/khan-exercises/issues/10957 for more.

Aaanyway, that regex probably worked back in 2012, but isn't working now
because many external script tags now have `data-main=...` between the
`script` and the `src=`. This fixes the regex so it can cope with that.

Test Plan:
Load /e/addition_1 in IE9 before this fix and see a 404 for local-only/requre.js
Load /e/addition_1 in IE9 after this fix and don't see any 404
Verify inline scripts still work by loading /e/logarithms_1 in IE9 and take all the hints, one of which depends on an inline script.

Reviewers: tom, alpert

Reviewed By: tom, alpert

Subscribers: alpert

Differential Revision: https://phabricator.khanacademy.org/D17664