Force new hints to receive focus so that they are read by a screen reader. Fixes T3748.
Summary: This changes the logic that was introduced by D16889 such that the hints are now inserted normally but they receive focus to make sure that the screen reader will read of the newly-introduced hint.

Test Plan: View an exercise with a screen reader. Click the "Show Hint" button. Note that the newly-introduced hint is immediately read off by the screen reader.

Reviewers: aria, jordan

Reviewed By: jordan

Subscribers: #a11y

Maniphest Tasks: T3748

Differential Revision: https://phabricator.khanacademy.org/D17990