Ensures that the "Show scratchpad" button works
Summary:
On some exercises (e.g. http://localhost:8080/math/pre-algebra/decimals-pre-alg/dividing-decimals-pre-alg/e/dividing_decimals_3),
there's an SVG used to store hints, but sometimes the SVG is so large that it flows over the "Show scratchpad" button.
Previously this meant that the "Show scratchpad" button couldn't be clicked, but I gave the button a z-index so it can be clicked.

Test Plan:
- Visit http://localhost:8080/math/pre-algebra/decimals-pre-alg/dividing-decimals-pre-alg/e/dividing_decimals_3
- Ensure that the "Show scratchpad" (and "Hide scratchpad") buttons work as expected

Reviewers: david

Reviewed By: david

Subscribers: alan

Differential Revision: https://phabricator.khanacademy.org/D18067