Pass up item ID along with the corretness of item when checking an answer.
Summary: Higher up in the tree when constructing attemptHistory, information will be available about which item the user actually got right and wrong instead of just an array of right/wrong values that can only be hackily matched to their original problem identities.

Test Plan: Tested in browser, and will be tested with mor examples when testing for skill checks.

Reviewers: riley, emily

Differential Revision: https://phabricator.khanacademy.org/D18426