Fix js errors in local mode for exercise screenshotting
Summary:
We have a nightly process that grabs screenshots of all of our exercises to use
in various places around the site. This process uses khan-exercises in local
mode (i.e., standalone, outside of webapp) so that it can pass in specific
seeds and problem types in order to screenshot particular problems.

For example, it does things like load the following URL and screenshot
the contents of `div#problemarea`:
http://sandcastle.kasandbox.org/media/castles/Khan:master/exercises/law_of_cosines.html?problem=find-side&seed=64

This diff fixes some javascript errors that occur only in local mode that are
currently preventing the screenshotter from doing its thing.

Test Plan:
1. From khan-exercises, run `python -m SimpleHTTPServer 9999`
2. Load http://localhost:9999/exercises/law_of_cosines.html?problem=find-side&seed=64
3. Verify there are no JS errors in the console
4. Copy this new interface.js into webapp/khan-exercises
5. Load http://localhost:8080/exercise/addition_1
6. Answer questions and verify in chrome network panel that POSTs are going to /api/internal/.../attempt
7. But mostly cross fingers this doesn't break anything in D18611 and D18675

Reviewers: phillip, johnsullivan

Reviewed By: phillip, johnsullivan

Differential Revision: https://phabricator.khanacademy.org/D18742