Avoid a denominator that imposes impossible constraints
Summary:
After 52d35b2eaabe109bf2880b83cd13aa7bb7ed3113, which improved the hints in this exercise, we found a number of bug reports that claimed mastery challenges could not be opened. After some further investigation, it came to be known that it was a problem with this exercise.

In particular, loading localhost:8000/exercises/comparing_fractions_1.html?debug&problem=same-denominator&seed=2 would not load.

This is because we have these #s:
DEN_1 = 2
DEN_2 = 2
NUM_1 = 1
NUM_2 needs to be a number between 1 and 0 that isn't 0... hmmm....difficult difficult lemon difficult.

Now we sidestep the DEN_2 possibility, and we hope that no more browsers hang in existential crisis in search of this impossible NUM_2.

Test Plan:
Load the aforementioned seed successfully
Cross fingers that nothing else breaks
Believe that everything will be completely fine since Eater was doing most of the thinking and Marcia was doing most of the typing!

Reviewers: eater

Reviewed By: eater

Differential Revision: https://phabricator.khanacademy.org/D19680